Breakout ecm_prep shared methods and variable classes in separate modules.
